refactor: move component definitions into new folder
