+++ medieval-recipe-app/src/components/ClickableSubstitution.jsx
